Ensures that files are always read in 16bit.

+++ dslr_scan_helper/lib.py
@@ -2,6 +2,7 @@ import numpy as np
 
 def rescale(array, old_min, old_max, new_min, new_max):
     result = array.astype(np.float32)
+
     result -= old_min
 
     result[result < 0] = 0
@@ -12,7 +13,9 @@ def rescale(array, old_min, old_max, new_min, new_max):
 
     result[result > new_max] = new_max
 
-    return result.astype(array.dtype)
+    result = result.astype(array.dtype)
+
+    return result
 
 def to8bit(img):
     return rescale(img, 0, 2**16, 0, 2**8).astype(np.uint8)
